end_of_lesson_78

+++ recipes/views.py
@@ -62,10 +62,24 @@ def recipe_update_view(request,id=None):
 
 @login_required
 def recipe_delete_view(request,id=None):
-    obj = get_object_or_404(Recipe,id=id,user=request.user)
-    success_url = reverse('recipes:list')
+    try:
+        obj = Recipe.objects.get(id=id,user=request.user)
+    except:
+        obj = None
+
+    if obj is None:
+        if request.htmx:
+            return HttpResponse("未找到")
+        raise Http404
+    
     if request.method=='POST':
         obj.delete()
+        success_url = reverse('recipes:list')
+        if request.htmx:
+            headers = {
+                'HX-Redirect':success_url
+            }
+            return HttpResponse("删除成功",headers=headers)
         return redirect(success_url)
     context = {
         'object':obj
@@ -74,10 +88,23 @@ def recipe_delete_view(request,id=None):
 
 @login_required
 def recipe_ingredient_delete_view(request,parent_id=None,id=None):
-    obj = get_object_or_404(RecipeIngredient,recipe__id=parent_id,id=id)
-    success_url = reverse('recipes:detail',kwargs={'id':parent_id})
+    try:
+        obj = RecipeIngredient.objects.get(recipe__id=parent_id,id=id,recipe__user=request.user)
+    except:
+        obj = None
+
+    if obj is None:
+        if request.htmx:
+            return HttpResponse("未找到")
+        raise Http404
+    
     if request.method=='POST':
+        name = obj.name
         obj.delete()
+        success_url = reverse('recipes:detail',kwargs={'id':parent_id})
+        if request.htmx:        
+            # return HttpResponse("<span style='color:#ccc;'>材料已删除</span>")
+            return render(request,'recipes/partials/ingredient-inline-delete-response.html',{"name":name})
         return redirect(success_url)
     context = {
         'object':obj
